# Make aws_instance_invalid_ami rule faster